requirements.txt
wheezy.web
pycrypto
gunicorn
psycopg2